Move IJsonHelper and associated types to Buildvana.Core.Json
